refactor: SSE 프록시 로직을 shared/api/sseProxy로 추출
